@@ -1,15 +1,14 @@
 package ch.hevs.apartments;
 
 import android.os.Bundle;
-import android.app.Activity;
 import android.content.Intent;
-import android.support.v7.app.ActionBar;
-import android.support.v7.app.AppCompatActivity;
-import android.support.v7.widget.Toolbar;
-import android.view.Menu;
+import androidx.appcompat.app.ActionBar;
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.appcompat.widget.Toolbar;
+
 import android.view.MenuItem;
-import android.widget.Toast;
-import android.support.v4.app.NavUtils;
+
+import androidx.core.app.NavUtils;
 
 public class About extends AppCompatActivity {
 	private String address;
